requirements.txt
django
django-extensions
django-safedelete
django-environ
djangorestframework
mysqlclient

celery[redis]
gevent
psutil

requests
pytest
pytest-django